Extract stats info to new directory
